Don't require user to set YSV_HOST
The original reason for having this required configuration parameter was that
I didn't think to read the hostname off the request. Doing it this way has the
additional benefit of a single server being able to serve a multitude of
hostnames.

Derek helped me realize this by repeating "there should be sane defaults" like
9000 times (maybe over 9000) till I got so tired of trying to make sure that he
understood why I did it that way to begin with that I looked at the problem
from a different perspective and like in a dream the solution came to me.

fixes #10.